full frame inference

+++ holodecml/transforms.py
@@ -190,12 +190,12 @@ class Normalize(object):
         image = sample['image']  # .astype(np.float32)
 
         if self.mode == "norm":
-            image -= image.min()
-            image /= image.max()
+            image -= image.min().astype(image.dtype)
+            image /= image.max().astype(image.dtype)
 
         if self.mode == "stan":
-            image -= image.mean()
-            image /= image.std()
+            image -= image.mean().astype(image.dtype)
+            image /= image.std().astype(image.dtype)
 
         if self.mode == "sym":
             image = -1 + 2.0*(image - image.min())/(image.max() - image.min())
